Add an EditorConfig (#2)
EditorConfig [^1] is a project that provides a common standard to
configure many IDEs and editors through a shared configuration file. The
goal is to help developers use a consistent coding style by configuring
their tools in the same way.

An EditorConfig with a few sensible defaults has been added to the
repository. The most controversial choice is probably the strict 80 line
length, which is a personal preference.