Add no-op status checks
Required status checks require that every status is reported with as a
"success". This doesn't work with path filters, which will not report a
status at all. The official recommendation[^1] is to create a no-op
status check that has the inverse of the path filter so that a status
gets reported.

[^1]: https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/defining-the-mergeability-of-pull-requests/troubleshooting-required-status-checks#handling-skipped-but-required-checks